html/semantics/embedded-content-0/media-elements/currentSrc_nonempty_after_setting_src.html

<!doctype html>
<html>
 <head>
  <title>{audio,video}.currentSrc - non-empty after setting src attribute</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
  <p><a href="http://dev.w3.org/html5/spec/Overview.html#dom-media-currentsrc">spec reference</a></p>
  <div id="log"></div>
  <script>
test(function() {
  var v = document.createElement("audio");
  assert_equals(
    v.currentSrc,
    "",
    "audioElement.currentSrc should be empty if there is no source");
  v.src = ".";
  var t = async_test("audioElement.currentSrc should not be empty after setting src attribute (even to a bogus value)");
  window.setTimeout(function() {
   t.step(function() {
    assert_not_equals(
      v.currentSrc,
      "")});
   t.done();
  }, 0);
}, "audio.currentSrc - non-empty after setting src attribute");

test(function() {
  var v = document.createElement("video");
  assert_equals(
    v.currentSrc,
    "",
    "videoElement.currentSrc should be empty if there is no source");
  v.src = ".";
  var t = async_test("videoElement.currentSrc should not be empty after setting src attribute (even to a bogus value)");
  window.setTimeout(function() {
   t.step(function() {
    assert_not_equals(
      v.currentSrc,
      "")});
   t.done();
  }, 0);
}, "video.currentSrc - non-empty after setting src attribute");
  </script>
 </body>
</html>
